edit profile backend

+++ app.js
@@ -360,8 +360,11 @@ app.get("/logout", (req, res) => {
 
 //DeleteAccount
 app.get("/deleteAccount", async (req, res, next) => {
-  const _id = jwt.verify(req.session.token, "MySecureKey");
-  const result = await User.findOneAndDelete({ _id: _id }); //we can also use email which is also stored in session
+  const currentUserId = await _.pick(
+    jwt.verify(req.session.token, "MySecureKey"),
+    ["_id"]
+  );
+  const result = await User.findOneAndDelete({ _id: currentUserId._id }); //we can also use email which is also stored in session
   console.log("Deleted::::");
   console.log(result);
   req.session.destroy();
@@ -369,7 +372,7 @@ app.get("/deleteAccount", async (req, res, next) => {
   res.redirect("/");
 });
 
-//IMAGE UPLOAD AND STORING IN DATABASE
+//IMAGE UPLOAD AND STORING IN DATABASE FOR POSTS
 const storage = multer.diskStorage({
   destination: (req, file, cb) => cb(null, "./static/usersPost/"),
   filename: async (req, file, cb) => {
@@ -380,6 +383,20 @@ const storage = multer.diskStorage({
     cb(null, currentUserId._id + file.originalname);
   },
 });
+
+//IMAGE UPLOAD AND STORING DATABASE FOR PROFILE
+const profileStorage = multer.diskStorage({
+  destination: (req, file, cb) => cb(null, "./static/profiles/"),
+  filename: async (req, file, cb) => {
+    const currentUserId = await _.pick(
+      jwt.verify(req.session.token, "MySecureKey"),
+      ["_id"]
+    );
+    cb(null, currentUserId._id + file.originalname);
+  },
+});
+
+//file filter for image file
 const fileFilter = (req, file, cb) => {
   if (
     file.mimetype === "image/jpeg" ||
@@ -392,8 +409,18 @@ const fileFilter = (req, file, cb) => {
   }
 };
 
+//UPLOAD FOR POSTS
 const upload = multer({
   storage: storage,
+  limits: {
+    fileSize: 1024 * 1024 * 5,
+  },
+  fileFilter: fileFilter,
+});
+
+//UPLOAD FOR PROFILE
+const uploadProfileImage = multer({
+  storage: profileStorage,
   limits: {
     fileSize: 1024 * 1024 * 5,
   },
@@ -471,6 +498,54 @@ app.get("/profile", async (req, res, next) => {
   res.status(200).render("profile.pug", profile);
   next();
 });
+
+//Updating the profile
+app.post(
+  "/profile",
+  uploadProfileImage.single("profileImg"),
+  async (req, res, next) => {
+    const currentUserId = await _.pick(
+      jwt.verify(req.session.token, "MySecureKey"),
+      ["_id"]
+    ); // getting current user id
+    let profileImg = (await User.findOne({ _id: currentUserId }, { photo: 1 }))
+      .photo;
+    if (req.file) {
+      profileImg =
+        "./static/profiles/" + currentUserId._id + req.file.originalname; // path for the stored profile images
+    }
+    const username = req.body.username;
+    const city = req.body.city;
+    const workplace = req.body.workplace;
+    const contactno = req.body.contactno;
+    var newValues = {
+      $set: {
+        username: username,
+        city: city,
+        companyname: workplace,
+        contact: contactno,
+        photo: profileImg,
+      },
+    };
+    let message = "";
+    const updation = await User.updateOne(
+      { _id: currentUserId },
+      newValues,
+      function (err, res) {
+        if (err) {
+          message = "Something Went Wrong";
+        }
+      }
+    );
+    const profile = await User.findOne({ _id: currentUserId._id });
+    if (message === "") {
+      message = "Profile Updated Successfully.";
+    }
+    profile.message = message;
+    res.status(200).render("profile.pug", profile);
+  }
+);
+
 app.post("/ajax/:action", async (req, res, next) => {
   console.log("inside ajax endpoint");
   const action = req.params.action;
